+++ src/main/java/POM/Success.java
@@ -1,6 +1,6 @@
 package POM;
 
-import Pages.SuperTestNG;
+import Utility.SuperTestNG;
 
 public class Success extends SuperTestNG {
